fix(inference): finish live goal and split streaming regressions
